Doors + Interactions
I made a bunch of doors and created a script to allow the player to interact with the doors.  One interaction should open the doors when the player gets near them.  The other interaction should allow the player to transition to the next level if "e" is pressed in front of the door.